requirements/common.txt
Flask==1.1.2
requests==2.25.1
opentelemetry-api==1.1.0
opentelemetry-exporter-jaeger==1.1.0
opentelemetry-sdk==1.1.0
opentelemetry-instrumentation-flask==0.20b0
opentelemetry-instrumentation-requests==0.20b0
prometheus-client

